+++ app/src/main/java/com/gymassistant/Models/Series.java
@@ -4,18 +4,18 @@ package com.gymassistant.Models;
  * Created by KamilH on 2016-05-02.
  */
 public class Series {
-    private int id;
+    private long id;
     private Training training;
-    private int trainingId;
+    private long trainingId;
     private Exercise exercise;
-    private int exerciseId;
+    private long exerciseId;
     private int order;
     private int repeat;
     private int weight;
 
     public Series(){}
 
-    public Series(Training training, int trainingId, Exercise exercise, int exerciseId, int order, int repeat, int weight) {
+    public Series(Training training, long trainingId, Exercise exercise, long exerciseId, int order, int repeat, int weight) {
         this.training = training;
         this.trainingId = trainingId;
         this.exercise = exercise;
@@ -25,11 +25,11 @@ public class Series {
         this.weight = weight;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
 
@@ -41,11 +41,11 @@ public class Series {
         this.training = training;
     }
 
-    public int getTrainingId() {
+    public long getTrainingId() {
         return trainingId;
     }
 
-    public void setTrainingId(int trainingId) {
+    public void setTrainingId(long trainingId) {
         this.trainingId = trainingId;
     }
 
@@ -57,11 +57,11 @@ public class Series {
         this.exercise = exercise;
     }
 
-    public int getExerciseId() {
+    public long getExerciseId() {
         return exerciseId;
     }
 
-    public void setExerciseId(int exerciseId) {
+    public void setExerciseId(long exerciseId) {
         this.exerciseId = exerciseId;
     }
